+++ source/ProcessManager.Orchestrations/Processes/BRS_026_028/BRS_028/V1/Activities/EnqueueActorMessagesActivity_Brs_028_V1.cs
@@ -43,13 +43,22 @@ internal class EnqueueActorMessagesActivity_Brs_028_V1(
             .GetAsync(input.InstanceId)
             .ConfigureAwait(false);
 
-        await EnqueueActorMessagesAsync(orchestrationInstance.Lifecycle.CreatedBy.Value, input).ConfigureAwait(false);
+        var orchestrationInstanceInput = orchestrationInstance.ParameterValue.AsType<RequestCalculatedWholesaleServicesInputV1>();
+
+        await EnqueueActorMessagesAsync(
+            orchestrationInstance.Lifecycle.CreatedBy.Value,
+            input,
+            orchestrationInstanceInput).ConfigureAwait(false);
     }
 
-    private Task EnqueueActorMessagesAsync(OperatingIdentity enqueuedBy, ActivityInput input)
+    private Task EnqueueActorMessagesAsync(
+        OperatingIdentity enqueuedBy,
+        ActivityInput input,
+        RequestCalculatedWholesaleServicesInputV1 requestInput)
     {
-        var requestInput = input.RequestInput;
-
+        var resolution = requestInput.Resolution != null
+            ? Resolution.FromName(requestInput.Resolution)
+            : null;
         var energySupplierNumber = requestInput.EnergySupplierNumber != null
             ? ActorNumber.Create(requestInput.EnergySupplierNumber)
             : null;
@@ -59,6 +68,13 @@ internal class EnqueueActorMessagesActivity_Brs_028_V1(
         var settlementVersion = requestInput.SettlementVersion != null
             ? SettlementVersion.FromName(requestInput.SettlementVersion)
             : null;
+        var chargeTypes = requestInput.ChargeTypes != null
+            ? requestInput.ChargeTypes.Select(
+                    ct => new RequestCalculatedWholesaleServicesAcceptedV1.AcceptedChargeType(
+                        ChargeType: ct.ChargeType != null ? ChargeType.FromName(ct.ChargeType) : null,
+                        ChargeCode: ct.ChargeCode))
+                .ToList()
+            : [];
 
         var acceptedData = new RequestCalculatedWholesaleServicesAcceptedV1(
             OriginalActorMessageId: requestInput.ActorMessageId,
@@ -68,13 +84,14 @@ internal class EnqueueActorMessagesActivity_Brs_028_V1(
             RequestedByActorNumber: ActorNumber.Create(requestInput.RequestedByActorNumber),
             RequestedByActorRole: ActorRole.FromName(requestInput.RequestedByActorRole),
             BusinessReason: BusinessReason.FromName(requestInput.BusinessReason),
+            Resolution: resolution,
             PeriodStart: InstantPattern.General.Parse(requestInput.PeriodStart).GetValueOrThrow().ToDateTimeOffset(),
             PeriodEnd: InstantPattern.General.Parse(requestInput.PeriodEnd!).GetValueOrThrow().ToDateTimeOffset(),
             GridAreas: requestInput.GridAreas,
             EnergySupplierNumber: energySupplierNumber,
             ChargeOwnerNumber: chargeOwnerNumber,
             SettlementVersion: settlementVersion,
-            ChargeTypes: requestInput.ChargeTypes ?? []);
+            ChargeTypes: chargeTypes);
 
         return _enqueueActorMessagesClient.EnqueueAsync(
             Orchestration_Brs_028_V1.UniqueName,
@@ -86,6 +103,5 @@ internal class EnqueueActorMessagesActivity_Brs_028_V1(
 
     public record ActivityInput(
         OrchestrationInstanceId InstanceId,
-        RequestCalculatedWholesaleServicesInputV1 RequestInput,
         Guid IdempotencyKey);
 }