+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -4,6 +4,15 @@
 
 - Add protobuf contract `CalculationEnqueueCompletedV1`
 
+## Version 0.18.4
+
+- Update NuGet package properties `RepositoryUrl`, `PackageReleaseNotes` and `PackageDescription`.
+
+## Version 0.18.3
+
+- Add `Resolution` to `RequestCalculatedWholesaleServicesAcceptedV1`.
+- Add `RequestCalculatedWholesaleServicesAcceptedV1.AcceptedChargeType` type for `RequestCalculatedWholesaleServicesAcceptedV1.ChargeTypes`.
+
 ## Version 0.18.2
 
 - Update `RequestCalculatedEnergyTimeSeriesRejectedV1` with missing properties.